remaking evaluation

+++ src/main/java/com/example/myTestApp/service/updatedevaluator/parser/Parser.java
@@ -4,6 +4,6 @@ import com.example.myTestApp.service.updatedevaluator.token.Token;
 import java.util.List;
 
 public interface Parser {
-    Integer evaluate(List<Token> lexemes);
+    Double evaluate(List<Token> lexemes);
     Integer getNumOfOperands(List<Token> lexemes);
 }
